Update os_generic to include TLS also general updates to the crash handling.

+++ tccengine/example_host.c
@@ -27,7 +27,7 @@ int ecaller()
 void PostFn( TCCEngine * tce )
 {
 	printf( "POST %p %p\n", tce->state, tce->backuptcc );
-	if( tce->backuptcc ) tcccrash_deltag( tce->backuptcc );
+	if( tce->backuptcc ) tcccrash_deltag( (intptr_t)tce->backuptcc );
 	tcccrash_symtcc( tce->filename, tce->state );
 }
